refactor: rename benchmark package -> core, benchmark.py -> execution.py
The central execution package was named 'benchmark', producing the
confusing triple path omnibenchmark.benchmark.benchmark.BenchmarkExecution
and overloading the word 'benchmark' (tool / package / module / class).

Give it clear semantics:
  omnibenchmark/benchmark/             -> omnibenchmark/core/
  omnibenchmark/core/benchmark.py      -> omnibenchmark/core/execution.py
  omnibenchmark.benchmark.benchmark.BenchmarkExecution
                                       -> omnibenchmark.core.execution.BenchmarkExecution

All 45 importers + test patch targets rewritten. The stale empty core/
(an earlier attempt at this same rename) is removed. model/benchmark.py
(the Pydantic models, a different file) is intentionally untouched.
