+++ omnibenchmark/cli/dashboard.py
@@ -8,8 +8,8 @@ from typing import Any, Dict, List
 
 import click
 
-from omnibenchmark.benchmark import BenchmarkExecution
-from omnibenchmark.benchmark.dashboard import create_bettr_dashboard
+from omnibenchmark.core import BenchmarkExecution
+from omnibenchmark.core.dashboard import create_bettr_dashboard
 from omnibenchmark.logging import logger
 from omnibenchmark.cli.error_formatting import pretty_print_parse_error
 from omnibenchmark.model.validation import BenchmarkParseError
