+++ omnibenchmark/cli/run.py
@@ -16,12 +16,12 @@ from omnibenchmark.backend.resolver import ModuleResolver
 from omnibenchmark.backend._manifest import write_run_manifest
 from omnibenchmark.backend._metadata import save_metadata
 from omnibenchmark.backend.snakemake import SnakemakeGenerator
-from omnibenchmark.benchmark import BenchmarkExecution
-from omnibenchmark.benchmark._paths import truncate_path_filename
+from omnibenchmark.core import BenchmarkExecution
+from omnibenchmark.core._paths import truncate_path_filename
 from omnibenchmark.model.params import Params
 from omnibenchmark.cli.error_formatting import pretty_print_parse_error
 from omnibenchmark.logging import logger
-from omnibenchmark.benchmark import populate_git_cache
+from omnibenchmark.core import populate_git_cache
 from omnibenchmark.model import SoftwareBackendEnum
 from omnibenchmark.model.resolved import ResolvedNode, TemplateContext
 from omnibenchmark.model.validation import BenchmarkParseError
