+++ omnibenchmark/core/__init__.py
@@ -1,5 +1,5 @@
-from .benchmark import BenchmarkExecution
-from .benchmark import BenchmarkExecution as Benchmark  # Compatibility alias
+from .execution import BenchmarkExecution
+from .execution import BenchmarkExecution as Benchmark  # Compatibility alias
 from ._node import BenchmarkNode
 from .validator import Validator
 from .prefetch import populate_git_cache
