+++ omnibenchmark/core/_dot.py
