+++ omnibenchmark/core/_graph.py
