+++ omnibenchmark/core/cite.py
@@ -4,13 +4,13 @@ import logging
 import yaml
 from typing import Dict, Optional, Any, List
 
-from omnibenchmark.benchmark.benchmark import BenchmarkExecution
-from omnibenchmark.benchmark.repository_utils import (
+from omnibenchmark.core.execution import BenchmarkExecution
+from omnibenchmark.core.repository_utils import (
     RepositoryManager,
     get_module_repository_info,
     resolve_module_repository,
 )
-from omnibenchmark.benchmark.metadata import (
+from omnibenchmark.core.metadata import (
     ValidationResult,
     ValidationIssue,
     ValidationSeverity,
